Setting global vars

+++ db/migrate/20210820231219_add_tax_rate_to_ruler_country.rb
@@ -1,5 +1,5 @@
 class AddTaxRateToRulerCountry < ActiveRecord::Migration[6.0]
   def change
-    add_column :ruler_countries, :tax_rate, :integer
+    add_column :ruler_countries, :tax_rate, :float
   end
 end
